decompose the cluster http client from the kubernetes data client
