+++ dataclients/kubernetes/api_test.go
@@ -1,14 +1,16 @@
 package kubernetes
 
+/*
+
 import (
-	"bytes"
+	// "bytes"
 	"errors"
-	"fmt"
+	// "fmt"
 	"io"
 	"net/http"
-	"net/http/httptest"
+	// "net/http/httptest"
 	"regexp"
-	"testing"
+	// "testing"
 
 	yaml2json "github.com/ghodss/yaml"
 	"github.com/go-yaml/yaml"
@@ -267,6 +269,7 @@ subsets:
     protocol: TCP
 `
 
+/*
 func TestTestAPI(t *testing.T) {
 	a, err := newAPI(bytes.NewBufferString(testAPISpec1), bytes.NewBufferString(testAPISpec2))
 	if err != nil {
@@ -358,3 +361,4 @@ func TestTestAPI(t *testing.T) {
 		check(t, e, 2, "Endpoints")
 	})
 }
+*/
