+++ dataclients/kubernetes/fixtures/routegroups/minimal_example.eskip
@@ -1,1 +1,3 @@
-kube__rg__default__myapp____0_0: Host("^(example.org)$") && PathSubtree("/") -> "";
+kube__rg__default__myapp____0_0:
+	Host("^(example.org)$") && PathSubtree("/") ->
+	<"http://10.2.9.103:7272", "http://10.2.9.104:7272">;
