+++ dataclients/kubernetes/fixtures/routegroups/minimal_example.yaml
@@ -8,9 +8,44 @@ spec:
   backends:
   - name: myapp
     type: service
-    serviceName: myapp-svc
+    serviceName: myapp
     servicePort: 80
   routes:
   - pathSubtree: /
     backends:
     - backendName: myapp
+---
+apiVersion: v1
+kind: Service
+metadata:
+  labels:
+    application: myapp
+  name: myapp
+spec:
+  clusterIP: 10.3.190.97
+  ports:
+  - name: main
+    port: 80
+    protocol: TCP
+    targetPort: 7272
+  selector:
+    application: myapp
+  type: ClusterIP
+---
+apiVersion: v1
+kind: Endpoints
+metadata:
+  labels:
+    application: myapp
+  name: myapp
+  namespace: default
+subsets:
+- addresses:
+  - ip: 10.2.9.103
+    nodeName: ip-10-153-130-98.eu-central-1.compute.internal
+  - ip: 10.2.9.104
+    nodeName: ip-10-153-130-99.eu-central-1.compute.internal
+  ports:
+  - name: main
+    port: 7272
+    protocol: TCP
