+++ dataclients/kubernetes/transformrg.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	log "github.com/sirupsen/logrus"
-
 	"github.com/zalando/skipper/eskip"
 )
